ints only no language issues with commas

+++ corptools/views.py
@@ -449,7 +449,7 @@ def fuel_levels(request):
         structure_tree.append(
             {'structure': s,
              'fuel_req': structure_hourly_fuel,
-             "current_blocks": hours*structure_hourly_fuel,
+             "current_blocks": int(hours*structure_hourly_fuel),
              "90day": max(int((structure_hourly_fuel*90*24)-(hours*structure_hourly_fuel)), 0)}
         )
